Internal: Group together editor tests
