+++ tests/playwright/sanity/editor.test.ts-snapshots/navigator-empty-dark-mode-linux.jpg
